First stab at using creating a functional bridge.
br0.sh will create tap0 and add it and en0 to bridge br0. This
has worked but not sure its the best algorithm. One thing I did
find surprising is that I had to add a second interface, tap0,
to make it work. I was expecting only the bridge and eno1 would
be needed and traffic originating internally would be directed
to br0 and the out eno1 and the reverse for incoming. So the
traffic flow would be:
```
app <-> br0 <-> en01 <-> LAN <-> Gateway <-> internet
```

But I couldn't get it to work so it seems traffic is doing:
```
app <-> tap0 <-> br0 <-> en01 <-> LAN <-> Gateway <-> internet
```


eno1-dynamic.2.sh is an iteration of eno1-dynamic.sh which
removes tap0 and br0 if they exist and then brings eno1
down and backup so that it uses dhcp to get its configuration.
Note: this isn't super reliable but has worked.