Use sendmsg instead of sendto.
Actaully, either can be used by changing some conditional compilation.
I did this because I wanted to experiment using sendmsg rather than
sendto as it will be more efficient to use scatter gather in most
instances.

I also added two initialization routines init_ether_arp and init_ethhdr.

Finally I added testing of the response to verify the result. I needed
this because if I ARP'd a non-existent node I was not getting an error
from the recvfrom.

Actually, I've added two TODO's one to test recvmsg and the other to
verify that the data in the ethernet frame is an struct ether_arp.