Add a bunch of debug and use ip everywhere.
I'm able to use test-af_packet to send an ARP request via eno1 and see
a response on my desktop. But when trying to use tap0 the send call returns
without error but nothing is captured by wireshark and there is no
response. A obvious error is that tap0 isn't UP.