Support macvlan sub-interfaces
Using macvlan sub-interfaces allows a simple bridge to be formed using
an existing physical interface. And allows that interface to appear on
the network just like the physical interface. This is exactly what I
want to do for creating a user-space network stack.

macvlan.sh can add/del macvlan sub-interfaces